Prevented shadow copy deletion when file is in memory

+++ SkyEditor.Core/IO/GenericFile.cs
@@ -1450,7 +1450,7 @@ namespace SkyEditor.Core.IO
                     }
 
                     // Delete the temporary file if shadow copy is enabled, the current I/O provider is not null, the physical filename exists, and the physical filename is different than the logical one
-                    if (EnableShadowCopy && CurrentIOProvider != null && !string.IsNullOrEmpty(PhysicalFilename) && CurrentIOProvider.FileExists(PhysicalFilename) && Filename != PhysicalFilename)
+                    if (EnableShadowCopy && InMemoryFile == null && CurrentIOProvider != null && !string.IsNullOrEmpty(PhysicalFilename) && CurrentIOProvider.FileExists(PhysicalFilename) && Filename != PhysicalFilename)
                     {
                         CurrentIOProvider.DeleteFile(PhysicalFilename);
                     }
